refactor(healthplatform): update agent commands to use healthplatform.Bundle()
- All cmd/ entry points use healthplatform.Bundle() instead of importing
  fx subpackages directly
- healthplatform.Diagnose() replaces direct core/impl imports
- Adds hostnameimpl.Module() where needed for the new forwarder dependency
