+++ cmd/agent/subcommands/diagnose/command.go
@@ -46,7 +46,7 @@ import (
 	workloadmetafx "github.com/DataDog/datadog-agent/comp/core/workloadmeta/fx"
 	workloadmetainit "github.com/DataDog/datadog-agent/comp/core/workloadmeta/init"
 	haagentfx "github.com/DataDog/datadog-agent/comp/haagent/fx"
-	healthplatformfx "github.com/DataDog/datadog-agent/comp/healthplatform/fx"
+	healthplatform "github.com/DataDog/datadog-agent/comp/healthplatform"
 	logscompressorfx "github.com/DataDog/datadog-agent/comp/serializer/logscompression/fx"
 	metricscompressorfx "github.com/DataDog/datadog-agent/comp/serializer/metricscompression/fx"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
@@ -120,7 +120,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				workloadfilterfx.Module(),
 				autodiscoveryimpl.Module(),
 				haagentfx.Module(),
-				healthplatformfx.Module(),
+				healthplatform.Bundle(),
 				hostnameimpl.Module(),
 				logscompressorfx.Module(),
 				metricscompressorfx.Module(),
