+++ cmd/agent/subcommands/flare/command.go
@@ -57,7 +57,7 @@ import (
 	workloadmetafx "github.com/DataDog/datadog-agent/comp/core/workloadmeta/fx"
 	workloadmetainit "github.com/DataDog/datadog-agent/comp/core/workloadmeta/init"
 	haagentfx "github.com/DataDog/datadog-agent/comp/haagent/fx"
-	healthplatformfx "github.com/DataDog/datadog-agent/comp/healthplatform/fx"
+	healthplatform "github.com/DataDog/datadog-agent/comp/healthplatform"
 	haagentmetadatafx "github.com/DataDog/datadog-agent/comp/metadata/haagent/fx"
 	"github.com/DataDog/datadog-agent/comp/metadata/host/hostimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/inventoryagentimpl"
@@ -157,7 +157,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				workloadfilterfx.Module(),
 				autodiscoveryimpl.Module(),
 				fx.Supply(option.None[collector.Component]()),
-				healthplatformfx.Module(),
+				healthplatform.Bundle(),
 				// We need inventoryagent to fill the status page generated by the flare.
 				inventoryagentimpl.Module(),
 				hostimpl.Module(),
